requirements/base.txt
asgiref==3.2.10
auth0-python==3.14.0
autopep8==1.5.5
boto3==1.17.20
botocore==1.20.20
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==3.3.1
deepdiff==3.3.0
defusedxml==0.6.0
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-db-logger==0.1.10
django-debug-toolbar==2.2
django-extensions==3.0.8
django-filter==2.4.0
django-jsonfield==1.4.0
django-multiselectfield==0.1.12
django-rest-auth==0.9.5
django-sendgrid-v5==0.9.1
django-ses==1.0.2
django-smtp-ssl==1.0
django-storages==1.10
djangorestframework==3.11.1
djangorestframework-api-key==2.0.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
flake8==3.8.4
Flask==1.1.2
future==0.18.2
gevent==20.5.1
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
jsonpickle==2.0.0
jsonschema==2.6.0
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.9
psycopg2-binary==2.8.4
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyJWT==1.7.1
pyOpenSSL==17.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.12.0
python-http-client==3.3.2
python3-openid==3.2.0
pytx==0.5.9
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rest-framework-auth0==0.6.4
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.4
sendgrid==6.6.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
termcolor==1.1.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
Werkzeug==0.16.0
whitenoise==5.1.0
zope.event==4.5.0
zope.interface==5.2.0
